Update README.md
I guess that you wanted to retain control. I thought I would try to see if you accept the concept of a external source issuing a rekey event. Given what happened yesterday, it is not clear to me that the core would accept the idea that user's need to be able to regain access to their identifier.